MaxAge support.
When reading stream or allstream, events are checked to see if they are expired. To do this check the stream metadata is fetched. This has a cost that has been mitigated by caching the stream metadata for a period of time (default is 1 minute). When an expired event has been encountered, a background task is queued to delete it. While the check is done in ReadOnlyEventStoreBase, the background task that does the deletion is done in EventStoreBase.