Convert build.cake to use dotnet core
- Modified build.cake to take advantage of dotnet core build tools.
- Changed nuget packaging to use the dotnet core way of building
packages.
- ILMerge is only available for the Net46 framework, not dotnet core
- Due to differences in how "dotnet pack" works with project.json
compared to "nuget pack" and nuspec files, the location of the
ILmerged dll had to be changed to the bin directory instead of
artifacts. I could not find a way to replace the DLL in the built
nuget package using the "files" propery in project.json.
- I chose to leave the PDB in it's own symbols project which is the
default for the new nuget packages.