Build SqlStreamStore on Linux
Inspired by the work of Yves from the build-on-linux branch.

The `DockerInstance` will be used in place of `LocalInstance` when the platform is detected as `Unix`.

This currently assumes that there is a MsSql server available on the
loopback address on the default MsSql port (1433).

If a user chooses to build on linux, the user can grab the MsSql Docker
image (https://hub.docker.com/r/microsoft/mssql-server-linux/) and start
it in the following manner

```
docker run -d -e 'ACCEPT_EULA=Y' -e 'SA_PASSWORD=SqlStreamSt0re' -p 1433:1433 microsoft/mssql-server-linux:latest
```

To build on linux the required dependencies are `Mono` and `Dotnet Core
2.0`.